[dsm] Changed the translation of the allocation *_L to always be *_D.
Since on 32-bit systems the sizes of double and long are the same but on 64-bit
systems they are different. DSM will read 4-bytes from the allocation file regardless
because the allocation is built on a 32-bit system but will interpret it incorrectly
as 8-bytes on a 64-bit system. Setting _L and _D to translate to the same type removes
this issue.